[feat #66] 프롬프트 서빙 구현
