+++ backend/requirements.txt
@@ -43,6 +43,8 @@ six==1.16.0
 sniffio==1.3.0
 SQLAlchemy==2.0.23
 starlette==0.32.0.post1
+tiktoken==0.5.2
+tokenizers==0.15.0
 tqdm==4.66.1
 typing_extensions==4.9.0
 urllib3==1.26.8
